components and other stuff

+++ SimpleUx/Attributes/RangeSliderAttribute.cs
@@ -6,10 +6,10 @@ namespace Overworld.Ux.Simple {
   /// Denotes a numeric field that should appear as a range slider in the UI
   /// </summary>
   public class RangeSliderAttribute : Attribute {
-    internal int _min;
-    internal int _max;
+    internal float _min;
+    internal float _max;
 
-    public RangeSliderAttribute(int min, int max) {
+    public RangeSliderAttribute(float min, float max) {
       _min = min;
       _max = max;
     }
